test: Make added-to-cart-modal e2e viewport aware
